Eval returns a list of requirements
Eval now returns a list of cell whose evaluation is required to finish
the computations. Right now those requirements are not used so we use a
singleton type to avoid passing around vectors. The requirements will be
used when we need to split spreadsheets (e.g. for big files or for
parralelism)